+++ R/print.simmr_output.R
@@ -11,6 +11,6 @@ print.simmr_output <-
 function(x,...) {
   print(x$input)
   cat('The input data has been run via simmr_mcmc and has produced ')
-  cat(nrow(x$output$BUGSoutput$sims.matrix),'iterations over',x$output$BUGSoutput$n.chains,'MCMC chains.')
+  cat(nrow(x$output[[1]]$BUGSoutput$sims.matrix),'iterations over',x$output[[1]]$BUGSoutput$n.chains,'MCMC chains.')
   cat('\n\n')
 }