+++ NEWS.md
@@ -2,6 +2,7 @@
 
   - Fixed a bug in combine_sources that stopped it working for multiple groups
   - Added in a load more tests to increase code coverage
+  - Implemented checkmate for error checking in simmr_load
 
 # simmr 0.4.2